refactor: migrate Double to BigDecimal across price and quantity fields

+++ src/main/kotlin/com/qkt/execution/Trade.kt
@@ -1,12 +1,13 @@
 package com.qkt.execution
 
 import com.qkt.common.Side
+import java.math.BigDecimal
 
 data class Trade(
     val orderId: String,
     val symbol: String,
-    val price: Double,
-    val quantity: Double,
+    val price: BigDecimal,
+    val quantity: BigDecimal,
     val side: Side,
     val timestamp: Long,
 )
